Build a k-d tree

Grow a kd-tree by iteratively partitioning along the sample median of each coordinate. Stop partitioning when the number of observatins in a node is less than \(4\log(n)\)

BuildKDTree(X, bounded = F, ...)

Arguments

X

data matrix with n rows and d columns

bounded

if bounded = TRUE, then creates an initial bounding box according to user input parameters:

option

Define the bounding box by the number of obs. (option = ndat) or quantiles (option = qt) of observations included.

q

number/quantile of points to be excluded in each end of a dimension. For example, if data is two-dimensional, option = "ndat" and qt = 50, then exclude 50 obs. in the lower and upper end of x-coordinate, then for the observations within, exclude 50 obs. in the lower and upper end of y-coordinate.

...

Additional parameters if bounded = TRUE

Value

build_tree returns two values.

kdtree

the k-d tree built from root node.

nd

The number of bounded regions at every level, starting at level 1.

Details

BuildKDTree constructs a kd-tree from data by iteratively partitioning along the sample median each coordinate. Each node in the kd-tree represents a region, and it consists of the following information:

depth

level of the region. The depth of root node is 1.

ndat

number of samples in this region.

low, up

lower and upper bound on region, if it exists.

lower, upper

lower and upper confidence bound for average density.

bounded

whether the region is bounded. A region is bounded if all of (low, up) exist.

leaf

whether the node is leaf or not. A node is a leaf if the number of observations is less than \(4 \log n\), where \(n\) is the total number of samples. A node is partitioned if it is not a leaf.

leftchild, rightchild

pointers to the left and right child of a node. leftchild points to the node representing the region of observations below median in partition dimension, and rightchild points to the node representing the region of observations above the median.

Examples

# build a k-d tree for standard normal samples
X <- matrix(rnorm(1000), ncol = 2)
kd <- BuildKDTree(X)