Select nodes in the histogram

Extract maximal (w.r.t. inclusion, i.e. tree order) nodes that are bounded and whose empirical density is within confidence bounds.

SelectNodes(node, B, ahat, n)

Arguments

node

A k-d tree with confidence thresholds, output from SetBounds.

B

A matrix.

ahat

A vector of confidence levels at each depth, starting at depth = 1. Output of the funtion ConfLevel

n

Total number of observations.

Value

A matrix of rectangles in the histogram of obs. in the input node. A region in the histogram is represented by a row, which contains the lower and upper bounds, empirical density, lower and upper confidence bounds of this node, number of obs. and the depth of the node.

Details

A bounded node is selected if the empirical density is within the confidence bounds. The empirical density is $$ h = \frac{n+1}{N |R|}, $$ where \(N\) is the number of obs., and \(n\) is the number of obs. in the region, and \(|R|\) is the volume of the region. Otherwise, if the node is not a leaf, move on to the left and right children of the node.