Adjacency matrix of regions in the Beta Tree

Computes the adjacency matrix of all of the regions in a Beta Tree histogram

compute_adjacency_mat(hist, d)

Arguments

hist

A matrix. Each row represents one region in the histogram. If the histogram is constructed for d dimensional data, then the first d columns contain the lower bound of each dimension, and the next d columns contain the upper bound. The 2d + 1 column contains the empirical density. hist can be the output of BuildHist function.

d

A numeric value of the data dimension.

Value

A matrix of size N*N, where N is the number of regions in the histogram, i.e. number of rows in the histogram.

Details

Two regions are neighbors if and only if the interval (defined by the boundary) in each dimension intersects. For example, let the two regions \(X\) defined by \([X_{i,\mathrm{low}}], X_{i,\mathrm{up}}]\) and \(Y\) defined by \([Y_{i,\mathrm{low}}], Y_{i,\mathrm{up}}]\) (for \(i=1,\ldots, d\)), where \([X_{i,\mathrm{low}}]\) and \(X_{i,\mathrm{up}}]\) are the lower and upper bounds in dimension \(i\) for the region \(X\). Then, \(X\) and \(Y\) are neighbors if and only if \([X_{i,\mathrm{low}}], X_{i,\mathrm{up}}] \cap [Y_{i,\mathrm{low}}], Y_{i,\mathrm{up}}] \neq \emptyset\) for every \(i=1,\ldots, d\).

Examples

X <-  matrix(rnorm(2000, 0, 1), ncol = 2)
rect <- BuildHist(X)
adj <- compute_adjacency_mat(hist = rect, d = 2)